requirements.txt
diffusers>=0.24.0
transformers>=4.35.0
torch>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
opencv-python>=4.8.0
scikit-image>=0.21.0
matplotlib>=3.7.0
# Optional but recommended for better results
# controlnet-aux>=0.0.7  # For better Canny edge detection
# git+https://github.com/huggingface/image_gen_aux  # For depth-anything 